src/agents/promotion-agent/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools]
packages = ["data"]
py-modules = ["register"]

[project]
name = "promotion-agent"
version = "0.1.0"
description = "Promotion Agent using NVIDIA NeMo Agent Toolkit for dynamic pricing"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "nvidia-nat-langchain==1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "pyright>=1.1.0",
]

# NAT plugin entry point - CRITICAL for function discovery
[project.entry-points."nat.plugins"]
promotion_agent = "register"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
# NAT imports and metaclass patterns are resolved when installed in own environment
reportMissingImports = "warning"
reportGeneralTypeIssues = "warning"
reportCallIssue = "warning"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
